Refactored for tty
