+++ src/commands/add/account.ts
@@ -1,7 +1,7 @@
 import { EnumType } from 'cliffy/command/mod.ts';
-import { Select } from 'cliffy/prompt/mod.ts';
 import { SupportedProviders } from '../../@providers/index.ts';
 import { BaseCommand, CommandHelper, GlobalOptions } from '../base-command.ts';
+import { Inputs } from '../common/inputs.ts';
 
 const providerType = new EnumType(Object.keys(SupportedProviders));
 
@@ -19,7 +19,7 @@ const AddAccountCommand = BaseCommand()
 async function add_account_action(options: AddAccountOptions, account_name?: string) {
   const command_helper = new CommandHelper(options);
 
-  const name = await command_helper.promptForStringInputs(
+  const name = await command_helper.resourceInputUtils.promptForStringInputs(
     {
       name: 'name',
       schema: {
@@ -36,14 +36,14 @@ async function add_account_action(options: AddAccountOptions, account_name?: str
   }
 
   const providerName = options.provider ||
-    (await Select.prompt({
+    (await Inputs.promptSelection({
       message: 'What provider will this account connect to?',
       options: Object.keys(SupportedProviders),
     }));
 
   const providerType = providerName as keyof typeof SupportedProviders;
 
-  const credentials = await command_helper.promptForCredentials(providerType);
+  const credentials = await command_helper.accountInputUtils.promptForCredentials(providerType);
   const account = new SupportedProviders[providerType](
     name,
     credentials as any,
