+++ src/commands/apply/datacenter.ts
@@ -30,7 +30,7 @@ async function apply_datacenter_action(options: ApplyDatacenterOptions, name: st
     const datacenter = await parseDatacenter(config_path);
 
     let graph = new CloudGraph();
-    const vars = await command_helper.promptForVariables(graph, datacenter.getVariables());
+    const vars = await command_helper.datacenterUtils.promptForVariables(graph, datacenter.getVariables());
     datacenter.setVariableValues(vars);
     graph = await datacenter.enrichGraph(graph);
 
@@ -40,18 +40,18 @@ async function apply_datacenter_action(options: ApplyDatacenterOptions, name: st
     }, command_helper.providerStore);
 
     pipeline.validate();
-    await command_helper.confirmPipeline(pipeline, options.autoApprove);
+    await command_helper.pipelineRenderer.confirmPipeline(pipeline, options.autoApprove);
 
     let interval: number | undefined = undefined;
     if (!options.verbose) {
       interval = setInterval(() => {
-        command_helper.renderPipeline(pipeline, { clear: true });
+        command_helper.pipelineRenderer.renderPipeline(pipeline, { clear: true });
       }, 1000 / cliSpinners.dots.frames.length);
     }
 
     let logger: Logger | undefined;
     if (options.verbose) {
-      command_helper.renderPipeline(pipeline);
+      command_helper.pipelineRenderer.renderPipeline(pipeline);
       logger = winston.createLogger({
         level: 'info',
         format: winston.format.printf(({ message }) => message),
@@ -59,7 +59,7 @@ async function apply_datacenter_action(options: ApplyDatacenterOptions, name: st
       });
     }
 
-    command_helper.applyDatacenter(name, datacenter, pipeline, logger)
+    command_helper.datacenterUtils.applyDatacenter(name, datacenter, pipeline, logger)
       .then(async () => {
         console.log(`Datacenter ${existingDatacenter ? 'updated' : 'created'} successfully`);
         if (datacenterEnvironments.length > 0) {
@@ -71,7 +71,7 @@ async function apply_datacenter_action(options: ApplyDatacenterOptions, name: st
             }, environmet.name);
           }
           console.log('Environments updated successfully');
-          command_helper.doneRenderingPipeline();
+          command_helper.pipelineRenderer.doneRenderingPipeline();
         }
       }).catch((err) => {
         console.error(err);
@@ -79,9 +79,9 @@ async function apply_datacenter_action(options: ApplyDatacenterOptions, name: st
       }).finally(async () => {
         if (interval) {
           clearInterval(interval);
-          await command_helper.saveDatacenter(name, datacenter, pipeline);
-          command_helper.renderPipeline(pipeline, { clear: !options.verbose, disableSpinner: true });
-          command_helper.doneRenderingPipeline();
+          await command_helper.datacenterUtils.saveDatacenter(name, datacenter, pipeline);
+          command_helper.pipelineRenderer.renderPipeline(pipeline, { clear: !options.verbose, disableSpinner: true });
+          command_helper.pipelineRenderer.doneRenderingPipeline();
         }
       });
   } catch (err: any) {
