+++ src/commands/deploy.ts
@@ -90,12 +90,12 @@ async function deploy_action(options: DeployOptions, tag_or_path: string): Promi
       }, command_helper.providerStore);
 
       pipeline.validate();
-      await command_helper.confirmPipeline(pipeline, options.autoApprove);
+      await command_helper.pipelineRenderer.confirmPipeline(pipeline, options.autoApprove);
 
       let interval: number;
       if (!options.verbose) {
         interval = setInterval(() => {
-          command_helper.renderPipeline(pipeline, {
+          command_helper.pipelineRenderer.renderPipeline(pipeline, {
             clear: true,
             message: `Deploying ${tag_or_path} to ${environmentRecord.name}`,
           });
@@ -104,7 +104,7 @@ async function deploy_action(options: DeployOptions, tag_or_path: string): Promi
 
       let logger: Logger | undefined;
       if (options.verbose) {
-        command_helper.renderPipeline(pipeline);
+        command_helper.pipelineRenderer.renderPipeline(pipeline);
         logger = winston.createLogger({
           level: 'info',
           format: winston.format.printf(({ message }) => message),
@@ -118,14 +118,13 @@ async function deploy_action(options: DeployOptions, tag_or_path: string): Promi
           logger,
         })
         .then(async () => {
-          await command_helper.saveEnvironment(
+          await command_helper.environmentUtils.saveEnvironment(
             datacenterRecord.name,
             environmentRecord.name,
-            datacenterRecord.config,
             environment,
             pipeline,
           );
-          command_helper.renderPipeline(pipeline, {
+          command_helper.pipelineRenderer.renderPipeline(pipeline, {
             clear: !options.verbose,
             disableSpinner: true,
             message: `Deploying ${tag_or_path} to ${environmentRecord.name}`,
@@ -133,14 +132,13 @@ async function deploy_action(options: DeployOptions, tag_or_path: string): Promi
           clearInterval(interval);
         })
         .catch(async (err) => {
-          await command_helper.saveEnvironment(
+          await command_helper.environmentUtils.saveEnvironment(
             datacenterRecord.name,
             environmentRecord.name,
-            datacenterRecord.config,
             environment,
             pipeline,
           );
-          command_helper.renderPipeline(pipeline, {
+          command_helper.pipelineRenderer.renderPipeline(pipeline, {
             clear: !options.verbose,
             disableSpinner: true,
             message: `Deploying ${tag_or_path} to ${environmentRecord.name}`,
