+++ src/environments/parser.ts
@@ -6,7 +6,6 @@ import { buildEnvironment, EnvironmentSchema } from './schema.ts';
 
 const DEFAULT_SCHEMA_VERSION = 'v1';
 const ajv = new Ajv2019({ strict: false, discriminator: true });
-const __dirname = new URL('.', import.meta.url).pathname;
 
 const environment_validator = ajv.compile<EnvironmentSchema>(EnvironmentSchemaContents.default);
